events: Move message relation types into separate module
